+++ semcore/errors/src/Error.jsx
@@ -4,11 +4,7 @@ import createComponent, { Root, Component, sstyled } from '@semcore/core';
 import { getIllustrationPath } from '@semcore/illustration';
 import style from './style/errors.shadow.css';
 
-const testEnv = process.env.NODE_ENV === 'test';
-
-const version = testEnv ? '3.0.0' : preval`module.exports = require('../package.json').version`;
-
-export const getIconPath = (name) => getIllustrationPath(name, 'errors', version);
+export const getIconPath = (name) => getIllustrationPath(name);
 
 class RootError extends Component {
   static displayName = 'Error';