+++ semcore/errors/CHANGELOG.md
@@ -2,6 +2,12 @@
 
 CHANGELOG.md standards are inspired by [keepachangelog.com](https://keepachangelog.com/en/1.0.0/).
 
+## [3.3.5] - 2022-09-27
+
+### Changed
+
+- Moved svg illustrations to `@semcore/illustration` component.
+
 ## [3.3.4] - 2022-09-20
 
 ### Changed